some style updates
