+++ src/Closed.js
@@ -1,12 +1,26 @@
+import styled from 'styled-components';
 import FullScreenHeight from './FullScreenHeight';
 
-const Closed = () => {
+const P = styled.p`
+  text-align: center;
+`;
+
+const Closed = ({ footerHeight }) => {
   return (
-    <>
-      <FullScreenHeight unsetBreakpoint="none" items="center" justify="center">
-        <p>countdown garden is closed 7AM - 9PM, please check back later</p>
-      </FullScreenHeight>
-    </>
+    <FullScreenHeight
+      unsetBreakpoint="none"
+      items="center"
+      justify="center"
+      direction="column"
+      addtlOffset={footerHeight}
+    >
+      <P>
+        <em>countdown garden is closed 7AM - 9PM</em>
+      </P>
+      <P>
+        <em>touch some grass and check back later</em>
+      </P>
+    </FullScreenHeight>
   );
 };
 export default Closed;
